Move zod infers to types.ts, Refactorings
