+++ packages/reassure-compare/src/output/console.ts
@@ -6,7 +6,7 @@ import {
   formatRenderCountChange,
   formatRenderDurationChange,
 } from '../utils/format';
-import type { PerformanceMetadata } from '../utils/validate';
+import type { PerformanceMetadata } from '../types';
 
 export function printToConsole(data: CompareResult) {
   // No need to log errors or warnings as these were be logged on the fly
