+++ packages/reassure-compare/src/utils/format.ts
@@ -1,5 +1,4 @@
-import type { CompareEntry } from '../types';
-import type { PerformanceMetadata } from './validate';
+import type { CompareEntry, PerformanceMetadata } from '../types';
 
 /**
  * Utility functions used for formatting data into strings
