+++ packages/reassure-compare/src/utils/validate.ts
@@ -1,20 +1,16 @@
 import { z } from 'zod';
 import { hasDuplicateValues } from './array';
 
-const performanceMetadataSchema = z.object({
+export const performanceMetadataSchema = z.object({
   branch: z.string().optional(),
   commitHash: z.string().optional(),
 });
 
-export type PerformanceMetadata = z.infer<typeof performanceMetadataSchema>;
-
-const performanceHeaderSchema = z.object({
+export const performanceHeaderSchema = z.object({
   metadata: performanceMetadataSchema,
 });
 
-export type PerformanceHeader = z.infer<typeof performanceHeaderSchema>;
-
-const performanceEntrySchema = z.object({
+export const performanceEntrySchema = z.object({
   /** Name of the test scenario */
   name: z.string(),
 
@@ -40,19 +36,17 @@ const performanceEntrySchema = z.object({
   counts: z.array(z.number()),
 });
 
-export type PerformanceEntry = z.infer<typeof performanceEntrySchema>;
-
-const perfFileSchema = z
+const performanceEntriesSchema = z
   .array(performanceEntrySchema)
   .refine((val) => !hasDuplicateValues(val.map((val) => val.name)), {
     message: `Your performance result file contains records with duplicated names.
     Please remove any non-unique names from your test suites and try again.`,
   });
 
-export function parseMetadata(metadata: unknown) {
-  return performanceHeaderSchema.parse(metadata);
+export function parseHeader(header: unknown) {
+  return performanceHeaderSchema.parse(header);
 }
 
-export function parsePerformanceEntries(perfEntry: unknown) {
-  return perfFileSchema.parse(perfEntry);
+export function parsePerformanceEntries(entries: unknown) {
+  return performanceEntriesSchema.parse(entries);
 }
